Handle failed montages

+++ tools/build-video-previews.sh
@@ -13,8 +13,11 @@ error() {
 # ffmpeg has various non-zero exit codes that do not actually mean it failed, so ignore them and always treat as success.
 # -y: overwrite the file created by lock.
 # Redirect stdin to /dev/null to prevent reading from this script's stdin.
+ff_status=0
 ff() {
-  ffmpeg -loglevel 0 -hide_banner -y "$@" </dev/null || return 0
+  ff_status=0
+  ffmpeg -loglevel 0 -hide_banner -y "$@" </dev/null || ff_status=$?
+  return 0
 }
 
 require_value() {
@@ -27,6 +30,7 @@ require_value() {
 # Tries to lock FD 9 and execute a command. If lock acquisition fails, returns with status 1.
 # If command fails, script is terminated.
 function exclusive_run_helper() {
+  # Use file descriptor instead of passing command so that we can call script functions.
   flock -xn 9 || return 1
   # Because this function is called followed by `||`, this command will not cause entire script to exit even with `set -e`.
   # Therefore, force it to exit the script if it fails.
@@ -41,9 +45,9 @@ exclusive_run() {
   local file="$1"
   shift
   if [ ! -f "$file" ]; then
-    subshell_error_code=0
-    exclusive_run_helper "$@" 9>"$file" || subshell_error_code=$?
-    if [[ $subshell_error_code -eq 0 ]]; then
+    lock_error_code=0
+    exclusive_run_helper "$@" 9>"$file" || lock_error_code=$?
+    if [[ $lock_error_code -eq 0 ]]; then
       exclusive_run_result=true
       return 0
     fi
@@ -121,14 +125,22 @@ for file in "$source_dir_abs"/**/*.{mp4,m4v}; do
 
   # Create montage.
   montage_dest="$output_dir_abs/$rel_path/montage.jpg"
+  nomontage="$output_dir_abs/$rel_path/.nomontage"
   # Check before continuing as temporary montage shot files would have already been deleted.
-  if [ -f "$montage_dest" ]; then
+  if [ -f "$nomontage" ] || [ -f "$montage_dest" ]; then
     continue
   fi
-  montage_granularity=200
+  # We want to get a shot every 2 seconds except for the first and last 2 seconds.
+  montage_granularity="$(bc -l <<<"scale=0; $duration / 2 - 2")"
+  if [[ $montage_granularity -le 0 ]]; then
+    # Video is too short for a montage, so don't create one.
+    touch "$nomontage"
+    continue
+  fi
   montage_shots=()
+  montage_failed=false
   for ((montage_shot_no = 0; montage_shot_no < montage_granularity; montage_shot_no++)); do
-    montage_shot_pos="$(bc -l <<<"scale=2; $duration * $montage_shot_no / $montage_granularity")"
+    montage_shot_pos="$((montage_shot_no * 2 + 2))"
     # Avoid using subdirectories that might cause race conditions when creating and deleting concurrently.
     montage_shot_dest="$output_dir_abs/$rel_path/montageshot${montage_shot_no}.jpg"
     montage_shots+=("$montage_shot_dest")
@@ -138,7 +150,17 @@ for file in "$source_dir_abs"/**/*.{mp4,m4v}; do
       -vframes 1 \
       -q:v 2 \
       "$montage_shot_dest"
+    if [[ "$(stat --printf="%s" "$montage_shot_dest")" -eq "0" ]]; then
+      # Montage shot failed to be created. Don't create montage and don't try again in future.
+      # Keep empty shot file so that no other concurrent scripts try to recreate it and which shots failed are known.
+      montage_failed=true
+      break
+    fi
   done
+  if [ "$montage_failed" = true ]; then
+    touch "$nomontage"
+    continue
+  fi
   exclusive_run "$montage_dest" convert "${montage_shots[@]}" +append -resize x120 "$montage_dest"
   if [ "$exclusive_run_result" = true ]; then
     rm -f "${montage_shots[@]}"
